+++ addOns/customreport/src/main/java/org/zaproxy/zap/extension/customreport/ExtensionCustomReport.java
@@ -183,6 +183,18 @@ public class ExtensionCustomReport extends ExtensionAdaptor {
         return alertDetailstab.wascid();
     }
 
+    public boolean attack() {
+        return alertDetailstab.attack();
+    }
+
+    public boolean param() {
+        return alertDetailstab.param();
+    }
+
+    public boolean evidence() {
+        return alertDetailstab.evidence();
+    }
+
     public boolean requestHeader() {
         return alertDetailstab.requestHeader();
     }