+++ addOns/customreport/src/main/java/org/zaproxy/zap/extension/customreport/ReportLastScan.java
@@ -84,6 +84,9 @@ public class ReportLastScan {
             Boolean reference,
             Boolean cweid,
             Boolean wascid,
+            Boolean attack,
+            Boolean param,
+            Boolean evidence,
             Boolean requestHeader,
             Boolean responseHeader,
             Boolean requestBody,
@@ -107,6 +110,9 @@ public class ReportLastScan {
                 reference,
                 cweid,
                 wascid,
+                attack,
+                param,
+                evidence,
                 requestHeader,
                 responseHeader,
                 requestBody,
@@ -126,6 +132,9 @@ public class ReportLastScan {
             Boolean reference,
             Boolean cweid,
             Boolean wascid,
+            Boolean attack,
+            Boolean param,
+            Boolean evidence,
             Boolean requestHeader,
             Boolean responseHeader,
             Boolean requestBody,
@@ -215,17 +224,17 @@ public class ReportLastScan {
                     report.append("<wascid>").append(alert.getWascId()).append("</wascid>\r\n");
                 }
                 report.append("  <uri>").append(replaceEntity(alert.getUri())).append("</uri>\r\n");
-                if (alert.getParam().length() > 0) {
+                if (alert.getParam().length() > 0 && param) {
                     report.append("<param>")
                             .append(replaceEntity(alert.getParam()))
                             .append("</param>\r\n");
                 }
-                if (alert.getAttack() != null && alert.getAttack().length() > 0) {
+                if (alert.getAttack() != null && alert.getAttack().length() > 0 && attack) {
                     report.append("<attack>")
                             .append(replaceEntity(alert.getAttack()))
                             .append("</attack>\r\n");
                 }
-                if (alert.getEvidence() != null && alert.getEvidence().length() > 0) {
+                if (alert.getEvidence() != null && alert.getEvidence().length() > 0 && evidence) {
                     report.append("<evidence>")
                             .append(replaceEntity(alert.getEvidence()))
                             .append("</evidence>\r\n");
@@ -357,6 +366,9 @@ public class ReportLastScan {
                                 extension.reference(),
                                 extension.cweid(),
                                 extension.wascid(),
+                                extension.attack(),
+                                extension.param(),
+                                extension.evidence(),
                                 extension.requestHeader(),
                                 extension.responseHeader(),
                                 extension.requestBody(),