+++ addOns/customreport/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Update minimum ZAP version to 2.9.0.
+- Add Evidence to report (Issue 6151).
+- Make Parameter and Attack fields optional.
 
 ## [5] - 2019-08-30